backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = "PDF Hero Backend"
authors = ["Anfer <anfer@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.109.0"
uvicorn = "^0.27.0"
pydantic = "^2.6.0"
python-multipart = "^0.0.9"
pillow = "^10.2.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.0.0"
pytest-asyncio = "^0.23.5"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
